Remove experimental code
